+++ libs/components/theme/projects/theme/src/icons/icon-manifest.module.ts
@@ -1,0 +1,10 @@
+import {
+  NgModule
+} from '@angular/core';
+
+/**
+ * @deprecated The `SkyThemeIconManifestService` no longer needs the `SkyThemeIconManifestModule`.
+ * The `SkyThemeIconManifestModule` can be removed from your project.
+ */
+@NgModule()
+export class SkyThemeIconManifestModule { }