+++ libs/components/theme/src/app/docs/button/button-docs.component.ts
@@ -1,9 +1,0 @@
-import {
-  Component
-} from '@angular/core';
-
-@Component({
-  selector: 'app-button-docs',
-  templateUrl: './button-docs.component.html'
-})
-export class ButtonDocsComponent { }